Validare + Roluri

+++ Controllers/CommentsController.cs
@@ -5,6 +5,7 @@ using RedditNet.CommentFolder;
 using RedditNet.DataLayerFolder;
 using RedditNet.Models.CommentModel;
 using RedditNet.Models.DatabaseModel;
+using RedditNet.PostFolder;
 using RedditNet.UtilityFolder;
 using System.Xml.Linq;
 
@@ -35,23 +36,45 @@ namespace RedditNet.Controllers
             {
                 DatabaseMapper mapper = new DatabaseMapper();
                 CommentThreadModel cm = mapper.toThreadComment(dbComment, subId);
-
+                
                 return View(cm);
             }
-
+            
             return View("Error");
         }
-
+        
+        [Authorize(Roles ="Regular,Moderator,Admin")]
         [HttpPost("{subId}/{postId}/comments/edit/{commentId}")]
         public IActionResult Edit(String subId, String postId, int commentId, CommentUpdateModel c)
         {
-            DatabaseComment? result = dbComments.updateComment(postId, commentId, c);  
+            DatabaseComment? result = dbComments.updateComment(postId, commentId, c);
+
             if (result != null)
-                return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId});
-
-            return BadRequest();
+            {
+                if (_userManager.GetUserId(User) == result.User.Id || User.IsInRole("Admin"))
+                {
+                    TempData["message"] = "Your comment has been modified";
+                    return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
+                }
+                else
+                {
+                    TempData["message"] = "You can't edit a comment that is not yours";
+                    return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
+                }
+            }
+            else
+            {
+                CommentUpdateModel cm = new CommentUpdateModel();
+                cm.PostId = postId;
+                cm.UserId = _userManager.GetUserId(User);
+                
+                cm.Id =commentId;
+                cm.SubId = subId;
+                return View(cm);
+            }
+            
         }
-
+        [Authorize(Roles = "Regular,Moderator,Admin")]
         [HttpGet("{subId}/{postId}/comments/edit/{commentId}", Name = "EditComment")]
         public IActionResult EditForm(String subId, String postId, int commentId)
         {
@@ -66,7 +89,13 @@ namespace RedditNet.Controllers
                 cm.Id = dbc.Id;
                 cm.SubId = subId;
 
-                return View("Edit", cm);
+                if (cm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                    return View("Edit", cm);
+                else
+                {
+                    TempData["message"] = "You can't edit a comment that is not yours";
+                    return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
+                }
             }
 
             return View("Error");
@@ -85,6 +114,7 @@ namespace RedditNet.Controllers
             cm.UserId = _userManager.GetUserId(User);
 
             return View("Create", cm);
+            
         }
 
         [HttpGet("{subId}/{postId}/comments")]
@@ -121,22 +151,50 @@ namespace RedditNet.Controllers
             CommentNode n = new CommentNode(0, commentId);
 
             DatabaseComment? dbComment = dbComments.createComment(n, comment, user);
-            Console.WriteLine(dbComment.Text);
+            /*Console.WriteLine(dbComment.Text);*/
+            
             if (dbComment != null)
+
+            {
+                TempData["message"] = "Yor comment has been posted";
+                Console.WriteLine(dbComment.Text);
                 return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
-
-            return BadRequest();
+             
+            }
+            else
+            {
+               CommentCreateModel c1=new CommentCreateModel();
+                
+                c1.PostId = postId;
+                c1.Parent = commentId;
+                c1.SubId = subId;
+                c1.UserId = _userManager.GetUserId(User);
+                return View(c1);
+            }
+                
         }
-
+        [Authorize(Roles = "Regular,Moderator,Admin")]
         [HttpPost("{subId}/{postId}/comments/delete/{commentId}")]
         public IActionResult Delete(String subId, String postId, int commentId, CommentDeleteModel c)
         {
-            if (dbComments.deleteComment(postId, commentId, c) == true)
+            DatabaseComment? dbc = dbComments.readComment(postId, commentId);
+            if (dbc.User.Id == _userManager.GetUserId(User) || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+            {
+                if (dbComments.deleteComment(postId, commentId, c) == true)
+                {
+                    TempData["message"] = "The comment has been deleted";
+                    return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
+                }
+                return BadRequest();
+            }
+            else
+            {
+                TempData["message"] = "You can't delete a comment ";
+
                 return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
-
-            return BadRequest();
+            }
         }
-
+        [Authorize(Roles = "Regular,Moderator,Admin")]
         [HttpGet("{subId}/{postId}/comments/delete/{commentId}", Name = "DeleteComment")]
         public IActionResult DeleteForm(String subId, String postId, int commentId)
         {
@@ -149,7 +207,16 @@ namespace RedditNet.Controllers
                 cm.Id = dbc.Id;
                 cm.SubId = subId;
 
-                return View("Delete", cm);
+                if (cm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+                {
+
+                    return View("Delete", cm);
+                }
+                else
+                {
+                    TempData["message"] = "You can't delete a comment that is not yours";
+                    return RedirectToRoute("ShowPostComments", new { subId = subId, postId = postId });
+                }
             }
 
             return View("Error");
